+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - change: do not raise error if failed to get client/server address info
 - fix: path handling error when har2case har file and cwd != ProjectRootDir
+- fix: missing list type for request body
 
 ## 3.0.13 (2020-06-17)